Stowing-main/requirements.txt
absl-py==1.4.0
addict==2.4.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
egl-probe==1.0.2
executing==1.2.0
fasteners==0.18
fastjsonschema==2.16.3
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.3
freetype-py==2.3.0
fvcore==0.1.5.post20221221
glfw==2.5.9
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h5py==3.8.0
hsluv==5.0.3
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.22.0
ipython==8.12.1
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lit==16.0.2
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
# mujoco-py==2.0.2.13
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
plotly==5.14.1
pooch==1.7.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pymeshfix==0.16.2
PyOpenGL==3.1.6
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
pysdf==0.1.8
python-dateutil==2.8.2
pytorch3d==0.3.0
pytz==2023.3
pyvista==0.38.5
PyYAML==6.0
pyzmq==25.0.2
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
scooby==0.7.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.1.0
# torch==2.0.0
# torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transforms3d==0.4.1
trimesh==3.21.5
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vispy==0.12.2
vtk==9.2.6
wcwidth==0.2.6
Werkzeug==2.3.3
widgetsnbextension==4.0.7
yacs==0.1.8
zipp==3.15.0
pytorch-minimize
readchar
scipy
h5py
matplotlib
pymeshfix
seaborn
vispy
trimesh
pysdf
transforms3d
imageio
mujoco-py==2.0.2.13
numpy
scikit-learn
